refactor(shared): consolidate form serialization and property configs (#779)
Move duplicated code from web and mobile apps to shared package:

- Add `buildFormData` function to `@volleykit/shared/api`
- Add property configurations (ASSIGNMENT_PROPERTIES, EXCHANGE_PROPERTIES,
  COMPENSATION_PROPERTIES, REFEREE_BACKUP_PROPERTIES) to shared package
- Update web-app to use shared buildFormData with web-specific CSRF handling
- Update mobile to use shared buildFormData with mobile-specific token handling
- Remove duplicated property arrays from both apps

This reduces code duplication and ensures both platforms use identical
API property configurations.

Closes #779
