+++ packages/shared/src/api/index.ts
@@ -3,6 +3,8 @@
  */
 
 export * from './client';
+export * from './form-serialization';
+export * from './property-configs';
 export * from './queryKeys';
 export * from './validation';
 // schema.ts is generated from OpenAPI spec - run `npm run generate:api` from web-app
